Struct actix_web_sql_identity::SqlIdentityBuilder

pub struct SqlIdentityBuilder { /* fields omitted */ }

Methods

impl SqlIdentityBuilder

pub fn new<T: Into<String>>(uri: T) -> SqlIdentityBuilder

Creates a new SqlIdentityBuilder that constructs a SqlIdentityPolicy

Arguments

uri - Database connection string

Example

use actix_web::App;
use actix_web::middleware::identity::IdentityService;
use actix_web_sql_identity::SqlIdentityBuilder;

// Create the identity policy
let policy = SqlIdentityBuilder::new("postgres://user:pass@host/database")
                .pool_size(5)
                .response_header("X-MY-RESPONSE")
                .finish()
                .expect("failed to open database");

let app = App::new().middleware(IdentityService::new(
    policy
));

pub fn response_header(self, hdr: &'static str) -> SqlIdentityBuilder

Change the response header when an identity is remembered

Arguments

hdr - Response header name to use

pub fn pool_size(self, count: usize) -> SqlIdentityBuilder

Change how many SQL connections are in each pool

Arguments

count - Number of connections per pool

pub fn finish(self) -> Result<SqlIdentityPolicy, Error>

Finish building this SQL identity policy. This will attempt to construct the a pool of connections to the database specified. The type of database is determined by the uri set. On success, a new SqlIdentityPolicy is returned, on failure an error is returned.

pub fn sqlite(self) -> SqlIdentityBuilder

Forces a SQLite identity policy to be created. This function does not normally need to be used, new will automatically determine the appropriate variant by parsing the connection string. This function exists if the parsing fails

pub fn mysql(self) -> SqlIdentityBuilder

Forces a MySQL identity policy to be created. This function does not normally need to be used, new will automatically determine the appropriate variant by parsing the connection string. This function exists if the parsing fails

pub fn postgresql(self) -> SqlIdentityBuilder

Forces a PostgreSQL identity policy to be created. This function does not normally need to be used, new will automatically determine the appropriate variant by parsing the connection string. This function exists if the parsing fails